+++ Manager/LocaleManager.php
@@ -1,0 +1,34 @@
+<?php
+
+namespace Pim\Bundle\MagentoConnectorBundle\Manager;
+
+use Pim\Bundle\CatalogBundle\Manager\LocaleManager as BaseLocaleManager;
+use Symfony\Bridge\Doctrine\RegistryInterface;
+
+/**
+ * Custom locale manager
+ *
+ * @author    Julien Sanchez <julien@akeneo.com>
+ * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
+ * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ */
+class LocaleManager extends BaseLocaleManager
+{
+    /**
+     * Get locale choices
+     * Allow to list locales in an array like array[<code>] = <label>
+     *
+     * @return string[]
+     */
+    public function getLocaleChoices()
+    {
+        $locales = $this->getLocales();
+
+        $choices = array();
+        foreach ($locales as $locale) {
+            $choices[$locale->getCode()] = $locale->getCode();
+        }
+
+        return $choices;
+    }
+}